[mlir][sparse] refactor utilities into transform/utils dir
Separates actual transformation files from supporting
utility files in the transforms directory. Includes a
bazel overlay fix for the build (as well as a bit of
cleanup of that file to be less verbose and more flexible).
